+++ src/modules/mouseHover/store/actionsMouseHover.js
@@ -1,6 +1,7 @@
 import {createGfiFeature} from "../../../api/gfi/getWmsFeaturesByMimeType";
 import {buffer} from "ol/extent";
 import Point from "ol/geom/Point";
+import rawLayerList from "@masterportal/masterportalapi/src/rawLayerList";
 
 export default {
     /**
@@ -30,9 +31,17 @@ export default {
             featuresAtPixel = [];
             commit("setHoverPosition", evt.coordinate);
 
+            if (Config.mouseHover.highlightOnHover) {
+                dispatch("Maps/removeHighlightFeature", "decrease", {root: true});
+            }
+
             // works for WebGL layers that are point layers
             map.forEachFeatureAtPixel(evt.pixel, (feature, layer) => {
                 if (layer?.getVisible()) {
+                    if (Config.mouseHover.highlightOnHover) {
+                        dispatch("highlightFeature", {feature, layer});
+                    }
+
                     if (feature.getProperties().features) {
                         feature.get("features").forEach(clusteredFeature => {
                             featuresAtPixel.push(createGfiFeature(
@@ -117,5 +126,42 @@ export default {
                 }
             });
         }
+    },
+
+    highlightFeature ({dispatch, state}, {feature, layer}) {
+        const layerId = layer.get("id"),
+            featureGeometryType = feature.getGeometry().getType(),
+            featureId = feature.getId(),
+            styleObj = featureGeometryType.toLowerCase().indexOf("polygon") > -1 ?
+                Config.mouseHover.highlightVectorRulesPolygon ?? state.highlightVectorRulesPolygon :
+                Config.mouseHover.highlightVectorRulesPointLine ?? state.highlightVectorRulesPointLine,
+            highlightObject = {
+                type: featureGeometryType === "Point" || featureGeometryType === "MultiPoint" ? "increase" : "highlightPolygon",
+                id: featureId,
+                layer: layer,
+                feature: feature,
+                scale: styleObj.image?.scale
+            },
+            rawLayer = rawLayerList.getLayerWhere({id: layerId});
+
+        if (featureGeometryType === "LineString") {
+            highlightObject.type = "highlightLine";
+        }
+        layer.id = layerId;
+        highlightObject.zoomLevel = styleObj.zoomLevel;
+        if (rawLayer && rawLayer.styleId) {
+            highlightObject.styleId = rawLayer.styleId;
+        }
+        else if (layer && layer.styleId) {
+            highlightObject.styleId = layer.styleId;
+        }
+
+        highlightObject.highlightStyle = {
+            fill: styleObj.fill,
+            stroke: styleObj.stroke,
+            image: styleObj.image
+        };
+
+        dispatch("Maps/highlightFeature", highlightObject, {root: true});
     }
 };